fix: cast bigint parameters in UpdateVODProgress
Postgres infers a parameter's type from its context. In
'CASE WHEN $3 <> 0', the untyped literal 0 resolved $3 to integer even
though total_bytes is BIGINT, so any file larger than 2GB overflowed:

  ERROR: value "5127667944" is out of range for type integer

The failure aborted the whole statement, so downloaded_bytes never advanced
either — progress stayed frozen for exactly the downloads big enough to
care about, while the database log filled with errors once every five
seconds per active download.

Cast both parameters explicitly. $2 did not strictly need it, since
assigning straight to a bigint column already pins the type, but leaving one
cast and not the other invites the same mistake back.

Introduced in the narrow progress statement added in #34. Audited the rest
of the package for the same shape: this was the only statement passing a
bigint parameter into a comparison rather than straight to a column.

Regression tests use the size from the reported failure and cover the upsert
path too; confirmed failing without the casts.

+++ pkg/database/vod_cache.go
@@ -91,10 +91,15 @@ func (m *DBManager) UpdateVODProgress(streamID string, downloaded, total int64) 
 	}
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
+	// $3 is cast explicitly. Postgres infers a parameter's type from its context,
+	// and comparing it against the untyped literal 0 resolves it to integer — so
+	// any file over 2GB overflowed and aborted the whole statement, leaving
+	// progress frozen for exactly the downloads long enough to care about. $2
+	// needs no cast because assigning straight to a bigint column already pins it.
 	_, err := m.db.ExecContext(ctx, `
         UPDATE vod_cache
-           SET downloaded_bytes = $2,
-               total_bytes = CASE WHEN $3 <> 0 THEN $3 ELSE total_bytes END,
+           SET downloaded_bytes = $2::bigint,
+               total_bytes = CASE WHEN $3::bigint <> 0 THEN $3::bigint ELSE total_bytes END,
                last_access = CURRENT_TIMESTAMP
          WHERE stream_id = $1`, streamID, downloaded, total)
 	return err
